solution for exam example
Exercise 1: Write a Python function that takes as input two lists L1 and L2 (of any possible length
N) which contain one digit integers (integers in the both the lists range from 0 to 9) and returns a
matrix that contains for each couple of indexes in the two lists i and j, the corresponding english
notation for the digits concatenated with an underscore in the middle. If both the digits have value 0
the english word for the greather digit in the two lists should be inserted alone instead.